Support phoenix_html 3.X
This is needed to upgrade to Phoenix 1.6.

The original project that we’ve forked is abandoned.

I need to change some tests to stop them comparing against the internal
representation of a Phoenix.HTML safe data structure, as the specifics
of that seem to have changed even though the rendered string is the
same.